combined authService and otpService into PubliceService
